refactor: Lotto 일급 콜렉션 파일 분리
